+++ src/router/changes.md
@@ -7,7 +7,9 @@
 
 ### To Do (Router)
 
-
+1. adjust priorities(who in controll opf defaults?) and default values of pathOptions ? like put default values into the RouteNode instead of router and matchPath ?
+2. force and reload (NavigationOptions) is more or less the same
+3. allow to pass additional arguments in promise chain, from parrent to childrensm here: `chain = Promise.all([chain, asyncFn])`
 
 ## RouteNode